+++ src/main/java/org/opentripplanner/transit/raptor/rangeraptor/workerlifecycle/LifeCycleEventPublisher.java
@@ -9,20 +9,20 @@ import java.util.function.IntConsumer;
  */
 public class LifeCycleEventPublisher {
     private final IntConsumer[] setupIterationListeners;
-    private final Runnable[] prepareForNextRoundListeners;
+    private final IntConsumer[] prepareForNextRoundListeners;
     private final Runnable[] transitsForRoundCompleteListeners;
     private final Runnable[] transfersForRoundCompleteListeners;
     private final Consumer<Boolean>[] roundCompleteListeners;
     private final Runnable[] iterationCompleteListeners;
 
     @SuppressWarnings("unchecked")
-    public LifeCycleEventPublisher(LifeCycleBuilder builder) {
-        this.setupIterationListeners = builder.setupIterationListeners.toArray(new IntConsumer[0]);
-        this.prepareForNextRoundListeners = builder.prepareForNextRoundListeners.toArray(new Runnable[0]);
-        this.transitsForRoundCompleteListeners = builder.transitsForRoundCompleteListeners.toArray(new Runnable[0]);
-        this.transfersForRoundCompleteListeners = builder.transfersForRoundCompleteListeners.toArray(new Runnable[0]);
-        this.roundCompleteListeners = builder.roundCompleteListeners.toArray(new Consumer[0]);
-        this.iterationCompleteListeners = builder.iterationCompleteListeners.toArray(new Runnable[0]);
+    public LifeCycleEventPublisher(LifeCycleSubscriptions subscriptions) {
+        this.setupIterationListeners = subscriptions.setupIterationListeners.toArray(new IntConsumer[0]);
+        this.prepareForNextRoundListeners = subscriptions.prepareForNextRoundListeners.toArray(new IntConsumer[0]);
+        this.transitsForRoundCompleteListeners = subscriptions.transitsForRoundCompleteListeners.toArray(new Runnable[0]);
+        this.transfersForRoundCompleteListeners = subscriptions.transfersForRoundCompleteListeners.toArray(new Runnable[0]);
+        this.roundCompleteListeners = subscriptions.roundCompleteListeners.toArray(new Consumer[0]);
+        this.iterationCompleteListeners = subscriptions.iterationCompleteListeners.toArray(new Runnable[0]);
     }
 
     /* Lifecycle methods invoked by the Range Raptor Worker */
@@ -33,9 +33,9 @@ public class LifeCycleEventPublisher {
         }
     }
 
-    public final void prepareForNextRound() {
-        for (Runnable it : prepareForNextRoundListeners) {
-            it.run();
+    public final void prepareForNextRound(final int round) {
+        for (IntConsumer it : prepareForNextRoundListeners) {
+            it.accept(round);
         }
     }