+++ docs/Changelog.md
@@ -19,6 +19,7 @@
 - New Transit search algorithm, Raptor, replaces the AStar for all transit searches. 
 - Ability to switch off the fare service(#2912).
 - Limit the transit service period(#2925).
+- Removed unwanted cost added for wait time between access and transit with RangeRaptor (#2927)
 
 ## Ported over from the 1.x
 - Add application/x-protobuf to accepted protobuf content-types (#2839)